requirements.txt
django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
python-dotenv==0.21.0
Pillow==10.0.0